move function to component to control state
